requirements-heroku.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
gunicorn==21.2.0
psycopg2-binary==2.9.7

Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3 